Update for Ruby 3.0, breaking changes
Ruby 3.0 will cause some errors in this gem
(https://github.com/zvory/csv-safe/pull/1).

I am using xxx's changes here to account for that, but this is a
breaking change. The gemspec file requires ruby 3.0 for this reason,
even though at time of writing 3.0 is not out yet.

Do not use this version of the gem if you are not on Ruby 3.