Add `csv` as a dependency
In Ruby 3.4, CSV-related functionality will be removed from the standard
library, meaning csv-safe will stop working, as it assumes that it'll be
available from there.

The functionality previously in the standard library has been extracted
into the `csv` gem. To that end, this adds the `csv` gem as a dependency
to ensure things keep working as expected in the future.

Additionally, this silences a warning that is produced in Ruby 3.3.1, as
detailed in zvory/csv-safe#19.